move object management interfaces
