+++ extensions/mssql/src/objectManagement/ui/userDialog.ts
@@ -4,14 +4,15 @@
  *--------------------------------------------------------------------------------------------*/
 import * as azdata from 'azdata';
 import { ObjectManagementDialogOptions } from './objectManagementDialogBase';
-import { IObjectManagementService, ObjectManagement } from 'mssql';
+import { IObjectManagementService } from 'mssql';
 import * as localizedConstants from '../localizedConstants';
 import { AlterUserDocUrl, CreateUserDocUrl } from '../constants';
 import { isValidSQLPassword } from '../utils';
 import { DefaultMaxTableRowCount } from '../../ui/dialogBase';
 import { PrincipalDialogBase } from './principalDialogBase';
+import { User, UserType, UserViewInfo } from '../interfaces';
 
-export class UserDialog extends PrincipalDialogBase<ObjectManagement.User, ObjectManagement.UserViewInfo> {
+export class UserDialog extends PrincipalDialogBase<User, UserViewInfo> {
 	private generalSection: azdata.GroupContainer;
 	private ownedSchemaSection: azdata.GroupContainer;
 	private membershipSection: azdata.GroupContainer;
@@ -45,7 +46,7 @@ export class UserDialog extends PrincipalDialogBase<ObjectManagement.User, Objec
 
 	protected override async validateInput(): Promise<string[]> {
 		const errors = await super.validateInput();
-		if (this.objectInfo.type === ObjectManagement.UserType.SqlAuthentication) {
+		if (this.objectInfo.type === UserType.SqlAuthentication) {
 			if (!this.objectInfo.password) {
 				errors.push(localizedConstants.PasswordCannotBeEmptyError);
 			}
@@ -56,7 +57,7 @@ export class UserDialog extends PrincipalDialogBase<ObjectManagement.User, Objec
 				&& (this.options.isNewObject || this.objectInfo.password !== this.originalObjectInfo.password)) {
 				errors.push(localizedConstants.InvalidPasswordError);
 			}
-		} else if (this.objectInfo.type === ObjectManagement.UserType.LoginMapped && !this.objectInfo.loginName) {
+		} else if (this.objectInfo.type === UserType.LoginMapped && !this.objectInfo.loginName) {
 			errors.push(localizedConstants.LoginNotSelectedError);
 		}
 		return errors;
@@ -148,18 +149,18 @@ export class UserDialog extends PrincipalDialogBase<ObjectManagement.User, Objec
 		this.removeItem(this.generalSection, this.confirmPasswordContainer);
 		this.removeItem(this.formContainer, this.advancedSection);
 		switch (this.objectInfo.type) {
-			case ObjectManagement.UserType.LoginMapped:
+			case UserType.LoginMapped:
 				this.addItem(this.generalSection, this.loginContainer);
 				break;
-			case ObjectManagement.UserType.AADAuthentication:
+			case UserType.AADAuthentication:
 				this.addItem(this.formContainer, this.advancedSection);
 				break;
-			case ObjectManagement.UserType.SqlAuthentication:
+			case UserType.SqlAuthentication:
 				this.addItem(this.generalSection, this.passwordContainer);
 				this.addItem(this.generalSection, this.confirmPasswordContainer);
 				this.addItem(this.formContainer, this.advancedSection);
 				break;
-			case ObjectManagement.UserType.WindowsUser:
+			case UserType.WindowsUser:
 				if (this.objectInfo.loginName) {
 					this.addItem(this.generalSection, this.loginContainer);
 				}
